Added ID verification before deleting an order

+++ src/main/java/org/warm4ik/hub/oms/service/OrderService.java
@@ -68,6 +68,9 @@ public class OrderService {
   @Transactional
   public void deleteOrderById(UUID orderId) {
 
+    if (!orderRepository.existsById(orderId)) {
+      throw new NotFoundException("Order с id" + orderId + " не найден и не может быть удалён.");
+    }
     orderRepository.deleteById(orderId);
   }
